feat: Add SnapSync stage configuration and integration
